Enables landscape orientation.
Note that I also removed a Framelayout that was containing a sole textview, the one that indicates "Level 3" or "Custom".  Seems to work fine without it, though the bottom half of the text was obstructed at first, so I converted the Textview to a Button. As long as the text is not too big for itself it displays fine.