Silences nullpointerexception on call to get shared prefs.
The true cause of this is a little mysterious, I can't replicate it.  In order to get here, the user first had to fail to connect to the billing manager. Then, the isPurchased() method gets called as a backup to check the app's local cache for purchases.  If the context can't be had here, let's just assume they haven't purchased.